workers/cpp/requirements.txt
click==7.1.1
dnspython==1.16.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
kafka-python==2.0.1
MarkupSafe==1.1.1
pika==1.1.0
pymongo==3.10.1
python-dotenv==0.13.0
Werkzeug==1.0.1
